Update arm_template_V2.json
The original statment (select * from Table) does not always returns a resultset with the columns in the specific order that the json model file is created. So, we need to use the exact same column order in our select statement here as the one that is used to create the *.cdm.json model file.